Remove custom tab navigator
createBottomTabNavigator in v3 doesn't support animations
https://react-navigation.canny.io/feature-requests/p/bring-back-animation-for-tabnavigator
