--- app/components/tab-bar/index.ts
@@ -1,1 +1,0 @@
-export * from "./tab-bar"
