+++ app/components/tab-icon/icons/ar.active@2x.png
