+++ app/components/tab-icon/icons/ar.inactive.png
