+++ app/components/tab-icon/icons/ar.inactive@2x.png
