+++ app/components/tab-icon/icons/ar.inactive@3x.png
