+++ app/components/tab-icon/icons/calendar.active@2x.png
