+++ app/components/tab-icon/icons/calendar.active@3x.png
