+++ app/components/tab-icon/icons/calendar.inactive.png
