+++ app/components/tab-icon/icons/calendar.inactive@2x.png
