+++ app/components/tab-icon/icons/calendar.inactive@3x.png
