+++ app/components/tab-icon/icons/info.active.png
