+++ app/components/tab-icon/icons/info.active@2x.png
