+++ app/components/tab-icon/icons/info.active@3x.png
