+++ app/components/tab-icon/icons/info.inactive.png
