+++ app/components/tab-icon/icons/info.inactive@2x.png
