+++ app/components/tab-icon/icons/info.inactive@3x.png
