+++ app/components/tab-icon/icons/map.active.png
