+++ app/components/tab-icon/icons/map.active@2x.png
