+++ app/components/tab-icon/icons/map.active@3x.png
