+++ app/components/tab-icon/icons/map.inactive.png
