+++ app/components/tab-icon/icons/map.inactive@2x.png
