+++ app/components/tab-icon/icons/map.inactive@3x.png
