+++ app/components/tab-icon/index.ts
@@ -1,0 +1,1 @@
+export * from "./tab-icon"
