+++ storybook/storybook-registry.ts
@@ -17,7 +17,6 @@ require("../app/components/travel-option/travel-option.story")
 require("../app/components/attraction/attraction.story")
 require("../app/components/rating/rating.story")
 require("../app/components/attractions-list/attractions-list.story")
-require("../app/components/tab-bar/tab-bar.story")
 require("../app/components/footer/footer.story")
 require("../app/components/social-button/social-button.story")
 require("../app/components/contact/contact.story")
